indices to long fix, remove pytorch-nightly dependency

+++ utils/foveation.py
@@ -274,13 +274,13 @@ def apply_gaussian_foveation_pyramid(image: torch.Tensor, foveation_params: dict
     pyramid = create_pyramid(image, None, levels=1+len(foveation_params["peripheral_rings"]), scale_factors=scale_factors)
     for i, ring in enumerate([foveation_params["fovea"], *foveation_params["peripheral_rings"]]):
         scale_factor = scale_factors[i]
-        target_indices = ring["target_indices"]
-        source_indices = torch.round((ring["mus"] - 0.5) / (scale_factor)).int()
+        target_indices = ring["target_indices"].long()
+        source_indices = torch.round((ring["mus"] - 0.5) / (scale_factor)).long()
         n_indices = source_indices.size(1)
-        batch_idx = torch.arange(b).view(b, 1, 1).expand(-1, c, n_indices)
-        channel_idx = torch.arange(c).view(1, c, 1).expand(b, c, n_indices)
-        x_idx = source_indices[:, :, 1].view(b, 1, n_indices).expand(b, c, n_indices)
-        y_idx = source_indices[:, :, 0].view(b, 1, n_indices).expand(b, c, n_indices)
+        batch_idx = torch.arange(b).view(b, 1, 1).expand(-1, c, n_indices).long()
+        channel_idx = torch.arange(c).view(1, c, 1).expand(b, c, n_indices).long()
+        x_idx = source_indices[:, :, 1].view(b, 1, n_indices).expand(b, c, n_indices).long()
+        y_idx = source_indices[:, :, 0].view(b, 1, n_indices).expand(b, c, n_indices).long()
         # foveated_image[
         #     :, :, target_indices[:, 0], target_indices[:, 1]
         # ] = pyramid[i][torch.arange(b).view(b, 1).expand(-1, source_indices.size(1)), :, source_indices[:, :, 1], source_indices[:, :, 0]].transpose(1, 2)  # pyramid[i].transpose(0, 1)[:, source_indices[:, :, 0], source_indices[:, :, 1]]
